Export: if no log entries found, export
the export process checks the export history table to determine the last successful export. It then checks if any changes have been made since, and if so, exports. Previous to this commit, if the query for log entries returned null (b/c the DB table was corrupt), no export was done. Now it will force an update.