Add ldap_only option
Currently, passwords will come out of the local database OR ldap.
This is sort of insane, but people might want it. So add an option
called ldap_only which makes passwords only come from LDAP.

This is also communicated to the client, which then hides the options
to edit passwords