Send Session Cookie only if secure (HTTPS), remove cookie on logout
The Session Cookie should only be sent via HTTPS,
so set the -secure flag on all ->cookie() calls.

On logout, remove the cookie entirely from the browser by
sending an expired cookie (as before), but with an empty value.

Without this change, the cookie is kept in the browser, and an
extra cookie without a name (but with the value) is created;
seen with firefox on a Debian system.